mistral-it-agent/requirements.txt
mistralai>=2.1.2
streamlit
python-dotenv
rich
pandas
scikit-learn
pytest
